Close OSC thumbs missing if file on mounted drive, #4480 (#5571)
This commit will add an "Include files on a mounted remote drive"
subordinate setting to the existing "Enable thumbnail preview" setting
to allow the generation of OSC thumbnails for files on remote drives.

This merely provides a user interface for the existing
enableThumbnailForRemoteFiles setting that currently can only be toggled
from terminal using the defaults command.

[Matt Svoboda] This was already implemented, but this commit adds use of @low_batt's text & localization support