Fix crashes with custom mpv OSD settings, #5727 (#5742)
This commit will:
- Change PlayerCore.openMainWindow to reset the mpv
icc-profile-auto option to false before loading a file
- Change VideoView.setICCProfile to set
MPV_RENDER_PARAM_ICC_PROFILE in the render context before setting the
mpv icc-profile-auto option to true

This corrects a crash that occurs when IINA is configured to display the
mpv OSD.

This commit will also correct the spelling of initialized in EventController.
This corrects a spelling error that typos is now detecting.